Update 1 (#1466)
* style(editorconfig): use LF line endings for C# files

Align .editorconfig with .gitattributes, which already pins
"*.cs text eol=lf": every .cs file in the Git index is stored with
LF line endings, and Git checks them out as LF.

Previously the two configs contradicted each other: .editorconfig
required CRLF, so on every save editors rewrote files as CRLF, and
Git then silently normalized them back to LF on staging. This caused
a pointless back-and-forth conversion on each edit/commit cycle.

LF is chosen as the single convention because:
- the repository is already 100% LF in the index, so no history
  renormalization is needed;
- the project is built and run in Linux-based Docker containers,
  where LF is native;
- modern .NET tooling (VS 2022, Rider, dotnet format) handles LF
  files on Windows without issues.

No code changes are involved. Working-tree files that still have
CRLF are considered clean by Git and will converge to LF as they
are edited and saved.

* feat(login): портировать логин-сервер под протокол клиента 3.5.0.3

Перенесены изменения AAEmu.Login из проекта 3.5.0.3 (AAC3500) по материалам
побайтового сравнения 1.2 -> 3.5 (diff-3503); патчи применены выборочно
с --include='AAEmu.Login/**':

- LoginClient: новые сигнатуры ACJoinResponsePacket (слоты персонажей 6->2,
  AdditionalData=22) и ACAuthResponsePacket под клиента 3.5;
- LoginSession: переработанная state-машина логина;
- CARequestAuthTencentPacketHandler: рабочая аутентификация через
  KoreaAuthFlowFactory вместо no-op заглушки;
- добавлены CARequestAuth_004_Packet и его обработчик;
- структуры пакетов C2L/L2C и опкоды (CLOffsets, LCOffsets, GLOffsets,
  LGOffsets) синхронизированы с клиентом 3.5.0.3;
- добавлены CHANGELOG.MD и Docs/LoginServer3503Port_ru.md с описанием порта.

Сознательно не перенесено (следующие шаги по PORTING-1.2-to-3.5.md):
AccountId uint->ulong в AAEmu.Commons (затрагивает Game-сервер), game-часть
протокола 3.5 (шифрование, GamePacket level 5, X2EnterWorld*).

Проверено: сборка AAEmu.Login, AAEmu.UnitTests и AAEmu.Login.IntegrationTests
без ошибок; 1076/1076 unit-тестов успешно.

* fix(login): address greptile review findings on the 3.5.0.3 port

- AfsValue.FromULong: read IsPreSelectCharacterPeriod from bit 32 to match
  ToULong; bit 16 overlapped the LSB of AdditionalData and silently
  corrupted the flag for odd AdditionalData values
- CARequestAuth_004_Packet: catch XmlException from XDocument.Parse so
  malformed XML from the network is logged and ignored instead of
  throwing on the connection worker
- CARequestAuthTencentPacketHandler: guard against null/whitespace
  packet.Account instead of relying on the null-forgiving operator
- ACAuthResponsePacket: document _wsk as a 3.5.x protocol-level constant
- Translate remaining Russian comments in AAEmu.Login to English
- Drop BOM from the touched .cs files to match repo conventions

* docs(login): add English port documentation, translate changelog

- Add Docs/LoginServer3503Port.md as the canonical English version of
  the login server port documentation, cross-linked with the Russian one
- Translate CHANGELOG.MD to English

English is the branch language for comments and documentation going
forward.

* fix(game): Phase1 - You can go to the lobby

* feat(game): update item serialization and holding container for 3.5.0.3 protocol

- Update item base serialization for ArcheAge 3.5.0.3 client:
  - add ChargeUseSkillTime field
  - adjust detail type lengths (Equipment 39, Mate 21, add ItemDetailType12)
  - move Durability, ChargeCount, ChargeTime, FreshnessTime,
    TemperPhysical, TemperMagical, RuneId and GemIds to base Item class
- Rework EquipItem detail read/write for 3.5.0.3 equipment data
  (EfainCubeId, LunaStone, RemainingExperience, EvolvedAttributeValues)
- Rename internal _holdingContainer field to public HoldingContainer
  property and update all usages across managers, packets and skills
- Register CSSecurityReportPacket in GameNetwork

* fix(game): update Ability

* feat(game): port 3.5.0.3 character packet fields and ability set model

- add HeirExp and HighAbilityRsc to Unit; write real values in
  Character.Write instead of zero stubs (heirExp, localLaborPower,
  crimeScore, forceNameChange, highAbilityRsc)
- extend SCUnitPointsPacket with highAbilityRsc and update all
  call sites (Unit, Character, Npc, Mate, Slave, TrialData,
  InstantGame, skill effects, regen task, resurrect, Heal/Revive)
- add character fields: LocalLaborPower, CrimeScore, ForceNameChange,
  HighAbility1-3, AbilitySets, AbilitySetFreeActivationCount,
  UsableAbilitySetSlotCount, ArenaDailyEntryCount
- add AbilitySetInfo model for saved ability/high-ability sets
- add CosplayVisual flag (0x20) to CharacterVisualOptions read/write
- add BattlePetSpell (6) to ActionSlotType
- make Actability a PacketMarshaler with Write (pisc Id/Point + Step)

* fix(game): enable Dwarf/Warborn character creation for 3.5.0.3 client

Replace the legacy 1.2 feature set (11 bytes) with the retail 3.5.0.3
fset (25 bytes) so the client unlocks Dwarf and Warborn races in the
character creation UI.

Also align related packets with the 3.5.0.3 protocol:
- SCInitialConfigPacket: write idleKickTime as int32, drop the 1.2-only
  secondPriceType field, and append the missing trailing fields
  (enable, pcbang, premium, maxch, honorPointDuringWarPercent, uccver,
  memberType)
- CSCreateCharacterPacket: read the trailing introZoneId (int32) field
- FeaturesManager: fix stale comment (the flag enables, not disables,
  Dwarf/Warborn creation)

* fix(game): port 3.5.0.3 character creation and appearance pipeline

- SCUnitStatePacket: port Inventory_Equip3 (validFlags-based equip)
- Write body part slots 19-25 as u32 templateId in character equip
  (char entry and unit state) as the 3.5.0.3 client expects
- CharacterManager.Create: validate ability and race/gender template,
  null-check equip item creation (no more disconnects on bad input)
- Apply template body part fallback only to Face/Hair/Body slots
- Fill ModelParams.ModelId from the race template when the client
  sends 0 so the lobby renders the correct base face model
- Map Horns/Stabilizer equipment slot types in EquipmentContainer
- Add debug logging of the create-character packet (plaintext and
  ciphertext) and model params round-trip

+++ AAEmu.Game/Core/Managers/CashShopManager.cs
@@ -215,7 +215,7 @@ public class CashShopManager(IWorldManager worldManager, IAccountManager account
     /// <param name="characterId"></param>
     /// <param name="shopItemId"></param>
     /// <returns>Resulting list of sales</returns>
-    public List<AuditIcsSale> GetSalesForShopItem(uint accountId, uint characterId, uint shopItemId)
+    public List<AuditIcsSale> GetSalesForShopItem(ulong accountId, uint characterId, uint shopItemId)
     {
         var res = new List<AuditIcsSale>();
 
@@ -262,8 +262,8 @@ public class CashShopManager(IWorldManager worldManager, IAccountManager account
         return res;
     }
 
-    public bool LogSale(uint buyerAccount, uint buyerChar,
-        uint targetAccount, uint targetChar,
+    public bool LogSale(ulong buyerAccount, uint buyerChar,
+        ulong targetAccount, uint targetChar,
         DateTime saleDate,
         uint shopItemId, uint sku,
         uint saleCost, CashShopCurrencyType saleCurrency,
